Support PHP-independent MCP tools in non-PHP IDEs
